Move OAuth2 server code to its own module
