+++ hq_superset/__init__.py
@@ -8,12 +8,14 @@ def flask_app_mutator(app):
     # Import the views (which assumes the app is initialized) here
     # return
     from superset.extensions import appbuilder
-    from . import api, hq_domain, views
+    from . import api, hq_domain, views, oauth2_server
 
     appbuilder.add_view(views.HQDatasourceView, 'Update HQ Datasource', menu_cond=lambda *_: False)
     appbuilder.add_view(views.SelectDomainView, 'Select a Domain', menu_cond=lambda *_: False)
     appbuilder.add_api(api.OAuth)
     appbuilder.add_api(api.DataSetChangeAPI)
+    oauth2_server.config_oauth2(app)
+
     app.before_request_funcs.setdefault(None, []).append(
         hq_domain.before_request_hook
     )
